Fix issue with _.isObject test before _.isFunction
If a serialize function is defined in the options object of a
LayoutManager, the test for _.isObject(function) happens first, which
passes, when in actuality the return value for the function is what
is wanted, not the function itself.

Doing the _.isFunction test before the _.isObject test fixes this.

Added two test cases which illustrate the fix.