Remove named function expression
In JScript, named function expressions generate two unique function
objects. From http://kangax.github.com/nfe/#jscript-bugs:

    var f = function g(){};
    f === g; // false

    f.expando = 'foo';
    g.expando; // undefined

This means that in Internet Explorer, the `afterRender` function passed
to `.on()` is distinct from the function passed to `.off()` when it
attempts to unbind itself.

This implementation is slightly more verbose, but it ensures correct
behavior in IE8.