Use documented jQuery API
Passing a string to the `$.fn.html` method works in the latest version
of the jQuery library, but this signature is neither documented [1] nor
tested [2]. This code does not work with the latest version of the
Cheerio library, and because it is non-standard, it is unlikely to be
implemented in a future version.

Use `$.fn.empty` and `$.fn.append` to implement the desired behavior in
a way that is compatable with Cheerio and unlikely to be removed in a
future release of jQuery.

[1] http://api.jquery.com/html/
[2] https://github.com/jquery/jquery/blob/2380028ec4a6a77401b867a51de26a3cb8e8d311/test/unit/manipulation.js